Consolidate HTTP/2 window_update increments
All increments for a single call to `Mint.HTTP2.stream` or recv are
sent in single frames per-stream plus the frame for the connection-level
increment. Any RST_STREAM frames and HEADER or DATA frames with the
End Stream flag set discard the stream-level window size increment.

This doesn't appear to have any effect in practice.